Update TME to take new inputs and output correct matrix element
Previously, a separate program had to be used to take the TME output
and actually get the matrix element. Updated TME to take new input
like what order of matrix element is being calculated, what carrier
is being captured, etc. to output the correct energy difference
for the delta function and and the correct matrix element.